main.c , Battery.c and battery.h error fixes. GPIO_Interface , GPIO_Program : removing the analog pins and follow up by changing joystick and ADC_Interface channel names.
ADC_Interface and ADC_Program : adding PRESCALER 0 to the enADC_Prescalers enum and changing ADC INIT function implementation

+++ HAL/BATTERY/Battery.h
@@ -13,7 +13,7 @@
 
 
 void batteryInit(void); // Initialize the battery monitoring system
-u16 readBatteryVoltage(void); // Read the current battery voltage and return it as a 16-bit 
+f32 readBatteryVoltage(void); // Read the current battery voltage and return it as a 16-bit 
 void checkBatteryVoltage(void);// Check the battery voltage and trigger an alert if it is below the defined threshold
 
 #endif
